cmd/segment-reaper: Move observer logic to a file
Move the observer implementation and the type definitions related with
it and helper functions to its own file.
